server/requirements.txt
openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git
fastapi>=0.115.0
uvicorn>=0.24.0
numpy>=1.19.0
openai>=1.0.0


